Armonizar códigos SINCO desde CMO o SINCO 2011

Esta función construye variables sinco1d, sinco2d, sinco3d y sinco4d a partir de códigos CMO (anteriores a 2013, o hasta el segundo trimestre de 2013) o directamente desde SINCO 2011 (desde el tercer trimestre de 2013 en adelante), utilizando una correspondencia incluida en el paquete y reglas manuales de clasificación cuando no hay equivalencia directa.

armoniza_sinco(data, codigos = NULL)

Arguments

data

Un data.frame con al menos anio, trimestre (como "t1", "t2", etc.) y p3coe. También puede contener sinco4d, pos_ocu, tue2.

codigos

(Opcional) data.frame de correspondencia CMO–SINCO. Si se omite, se usa la tabla interna del paquete.

Value

Un data.frame con variables armonizadas: cmo_4d, sinco4d, sinco3d, sinco2d, sinco1d.

See also

Examples

datos <- data.frame(
  anio = c(2011, 2012, 2013),
  trim = c("t1", "t2", "t3"),
  p3coe = c(7121, 4190, 1110),
  pos_ocu = c(1, 2, 1),
  tue2 = c(1, 2, 1)
)
armoniza_sinco(datos)
#>   anio trim p3coe pos_ocu tue2 cmo_4d sinco4d sinco3d sinco2d sinco1d
#> 1 2011   t1  7121       1    1   7121      NA      NA      NA       8
#> 2 2012   t2  4190       2    2   4190    6999     699      69       9
#> 3 2013   t3  1110       1    1   1110    2211     221      22      NA
#>   aplicar_reglas  cmo str_cmo cmo2d cmo3d
#> 1           TRUE 7121    7121    71   712
#> 2           TRUE 4190    4190    41   419
#> 3          FALSE   NA    <NA>    NA    NA