Procesar variables de análisis laboral y desajuste educativo

Esta función genera variables clasificatorias relacionadas con la ocupación, nivel educativo y desajuste entre ambos, a partir de los códigos de ocupación (p3coe) y nivel educativo alcanzado (cs_p13_1). Internamente armoniza los códigos SINCO (1, 2, 3 y 4 dígitos) usando correspondencias con códigos CMO y reglas auxiliares. También clasifica el nivel de habilidad de la ocupación, el nivel educativo alcanzado y construye indicadores de sobrecalificación, ajuste o subcalificación laboral.

procesar_vars_laborales(data)

Arguments

data

Un data.frame con variables como:

anio, trimestre: año y trimestre de la entrevista

coe_tipo: tipo de cuestionario ("ampliado" o "basico")

p3coe: código ocupacional

cs_p13_1, anios_es: nivel educativo alcanzado en categorías o años

clase2: clase de actividad económica

pos_ocu, tue2: posición en la ocupación y tipo de unidad económica

p2_4: experiencia laboral previa

p3i, p3j, p3j1, p3k1: variables sobre tipo de contrato

Value

Un data.frame con las variables originales y nuevas columnas:

sinco1d, sinco2d, sinco3d, sinco4d

skill_level, skill_actual

mismatch, mismatch2

nunca_trabajo, status_seq

contrato0, contrato1, temporal, temporal_seq

Details

Además, genera variables relacionadas con la experiencia previa (nunca_trabajo), estatus laboral combinado (status_seq) y características contractuales (contrato0, contrato1, temporal, temporal_seq) según el tipo de cuestionario.

See also